Feature/btrv2 integration (#554)
* adjusted deploy script to work with btr v2

* tested butter page with btrv2, changed info texts, added minor improvements in various places on the butter page

* added staking apy to est. apy, added tooltip to est apy, changed selectedToken to use keys, mint and redeem modal are now more similiar

* Update packages/app/pages/butter.tsx


* Refactor unneccessary useSWRs into useMemos.

* Always enable approve button if user isn't approved.

* Delete Testpage

* updated infoModal copy

* Update packages/app/components/BatchButter/MintRedeemToggle.tsx


* Update packages/app/components/BatchButter/MintRedeemToggle.tsx


* fixed changing chains on butter

+++ packages/hardhat/contracts/core/defi/butter/ButterBatchProcessingZapper.sol
@@ -120,7 +120,7 @@ contract ButterBatchProcessingZapper {
    */
   function claimAndSwapToStable(
     bytes32 _batchId,
-    uint256 _stableCoinIndex,
+    uint8 _stableCoinIndex,
     uint256 _min_amount
   ) external {
     //We can only deposit 3CRV which come from mintBatches otherwise this could claim Butter which we cant process here
